renaming: Move sub components of manager under manager
