+++ manager/clusterapi/node.go
@@ -2,7 +2,7 @@ package clusterapi
 
 import (
 	"github.com/docker/swarm-v2/api"
-	"github.com/docker/swarm-v2/state"
+	"github.com/docker/swarm-v2/manager/state"
 	"golang.org/x/net/context"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
