+++ manager/clusterapi/server_test.go
@@ -10,7 +10,7 @@ import (
 	"google.golang.org/grpc"
 
 	"github.com/docker/swarm-v2/api"
-	"github.com/docker/swarm-v2/state"
+	"github.com/docker/swarm-v2/manager/state"
 	"github.com/stretchr/testify/assert"
 )
