+++ manager/dispatcher/dispatcher.go
@@ -8,9 +8,9 @@ import (
 
 	log "github.com/Sirupsen/logrus"
 	"github.com/docker/swarm-v2/api"
-	"github.com/docker/swarm-v2/dispatcher/heartbeat"
 	"github.com/docker/swarm-v2/identity"
-	"github.com/docker/swarm-v2/state"
+	"github.com/docker/swarm-v2/manager/dispatcher/heartbeat"
+	"github.com/docker/swarm-v2/manager/state"
 	"golang.org/x/net/context"
 )
