+++ manager/orchestrator/orchestrator_test.go
@@ -5,8 +5,8 @@ import (
 	"time"
 
 	"github.com/docker/swarm-v2/api"
-	"github.com/docker/swarm-v2/state"
-	"github.com/docker/swarm-v2/state/watch"
+	"github.com/docker/swarm-v2/manager/state"
+	"github.com/docker/swarm-v2/manager/state/watch"
 	"github.com/stretchr/testify/assert"
 )
