+++ manager/state/doc.go
