+++ manager/state/memory_test.go
