+++ manager/state/raft.go
