+++ manager/state/raft_test.go
