+++ manager/state/util.go
