+++ manager/state/watch.go
@@ -2,7 +2,7 @@ package state
 
 import (
 	"github.com/docker/swarm-v2/api"
-	"github.com/docker/swarm-v2/state/watch"
+	"github.com/docker/swarm-v2/manager/state/watch"
 )
 
 // Event is the type used for events passed over watcher channels, and also
