+++ manager/state/watch/pubsub.go
