+++ manager/state/watch/pubsub_test.go
